requirements.txt
python-telegram-bot==21.1.1
google-genai
pillow
python-dotenv
